refactor: move updatePriceEvents to its own file
