+++ src/game-logic/reducers/index.test.js
@@ -42,28 +42,6 @@ jest.mock('../../constants', () => ({
   PRECIPITATION_CHANCE: 0,
 }))
 
-describe('updatePriceEvents', () => {
-  test('updates price events', () => {
-    const { priceCrashes, priceSurges } = fn.updatePriceEvents({
-      priceCrashes: {
-        'sample-crop-1': { itemId: 'sample-crop-1', daysRemaining: 1 },
-        'sample-crop-2': { itemId: 'sample-crop-2', daysRemaining: 3 },
-      },
-      priceSurges: {
-        'sample-crop-3': { itemId: 'sample-crop-3', daysRemaining: 5 },
-      },
-    })
-
-    expect(priceCrashes).toEqual({
-      'sample-crop-2': { itemId: 'sample-crop-2', daysRemaining: 2 },
-    })
-
-    expect(priceSurges).toEqual({
-      'sample-crop-3': { itemId: 'sample-crop-3', daysRemaining: 4 },
-    })
-  })
-})
-
 describe('updateFinancialRecords', () => {
   test('updates financial records', () => {
     const {
