+++ src/game-logic/reducers/updatePriceEvents.test.js
@@ -1,0 +1,23 @@
+import { updatePriceEvents } from './updatePriceEvents'
+
+describe('updatePriceEvents', () => {
+  test('updates price events', () => {
+    const { priceCrashes, priceSurges } = updatePriceEvents({
+      priceCrashes: {
+        'sample-crop-1': { itemId: 'sample-crop-1', daysRemaining: 1 },
+        'sample-crop-2': { itemId: 'sample-crop-2', daysRemaining: 3 },
+      },
+      priceSurges: {
+        'sample-crop-3': { itemId: 'sample-crop-3', daysRemaining: 5 },
+      },
+    })
+
+    expect(priceCrashes).toEqual({
+      'sample-crop-2': { itemId: 'sample-crop-2', daysRemaining: 2 },
+    })
+
+    expect(priceSurges).toEqual({
+      'sample-crop-3': { itemId: 'sample-crop-3', daysRemaining: 4 },
+    })
+  })
+})
